Update techspec.md
Made multiple grammatical, form and stylistic modifications. Added clarifying language about private networks, made specific claims on line 146 for Binh to validate. And Binh — need to ask if we are using the term "butterfly key" here but with no explanation of what that is. Thoughts? (The reference in the endnotes is not descriptive of the approach.)