Edit request from ID, per Frank
So this (much longer) whitepaper now seems to be the original whitepaper, with the original protocol spec added to it. ? ... My first impression is that the two were beneficial as separate documents. ...  Note we still have a link at the bottom of this whitepaper to the protocol spec, which presumably would now contain new / different content than that which was moved to this whitepaper? (We should ideally avoid repeating content across multiple documents - difficult to keep it in sync and potentially confusing / repetitive for readers.)